// enums/Event.d.ts
export declare enum Event {
    GameStart = 0,
    Kickoff = 1,
    HalfTime = 2,
    GameEnd = 3,
    Injury = 4,
    Goal = 5,
    Save = 6,
    Block = 7,
    Substitution = 8,
    Possession = 9,
    Defence = 10,
    EventLess = 11,
    Advance = 12,
    Retreat = 13,
    Corner = 14,
    FreeKick = 15
}